VOXELUTIL: fixed shadowing warning

+++ src/modules/io/tests/BufferedReadWriteStreamTest.cpp
@@ -294,8 +294,8 @@ TEST(BufferedReadWriteStreamTest, testUTF16) {
 	BufferedReadWriteStream stream;
 	core::String str = "foobar string %&";
 	EXPECT_TRUE(stream.writeUTF16BE(str));
-	const int utf16Len = stream.pos();
-	EXPECT_EQ(utf16Len, str.size() * sizeof(uint16_t));
+	const int64_t utf16Len = stream.pos();
+	EXPECT_EQ(utf16Len, (int64_t)(str.size() * sizeof(uint16_t)));
 	stream.seek(0);
 	core::String str2;
 	EXPECT_TRUE(stream.readUTF16BE(str.size(), str2));
